Fix datatype

+++ fmm.py
@@ -8,7 +8,7 @@ from scipy.fftpack import fft
 from scipy.sparse import diags
 
 
-def fmm1d_te_layer_modes(perm, period, k_0, k_x, N):
+def fmm1d_te_layer_modes(perm, period, k_0, k_x, N, dtype=np.complex128):
     '''Calculates the TE eigenmodes of a one-dimensional grating layer.
 
     Arguments
@@ -34,8 +34,9 @@ def fmm1d_te_layer_modes(perm, period, k_0, k_x, N):
     '''
     # number of points in x direction
     N_x = perm.size
+    perm = perm.astype(dtype)
     # Fourier coefficients of the permittivity
-    perm_fc = fft(perm) / (N_x - 1)
+    perm_fc = (fft(perm) / (N_x - 1)).astype(dtype)
 
     # take the first 2 * N positive and 0 frequency
     perm_fc_pos = perm_fc[:2 * N + 1]
@@ -44,21 +45,21 @@ def fmm1d_te_layer_modes(perm, period, k_0, k_x, N):
                                   perm_fc[-(2 * N):][::-1]), axis=0)
 
     # calculate grating
-    Gm = np.arange(-N, N + 1, 1) * 2 * np.pi / period
+    Gm = np.arange(-N, N + 1, 1, dtype=dtype) * 2 * np.pi / period
 
     # create the Toeplitz matrix containing the Fourier coefficients of perm
-    eps_hat = toeplitz(perm_fc_pos, perm_fc_neg)
+    eps_hat = toeplitz(perm_fc_pos, perm_fc_neg).astype(dtype)
     # create \hat K Matrix
-    K_hat_square = diags((Gm + k_x) ** 2, offsets=0).todense()
+    K_hat_square = diags((Gm + k_x) ** 2, offsets=0).todense().astype(dtype)
 
     # create final matrix
-    M_hat = (k_0 ** 2 * eps_hat - K_hat_square)
+    M_hat = (k_0 ** 2 * eps_hat - K_hat_square).astype(dtype)
 
     # calculate the eigenvalues and eigenvectors of M_hat
     eig_values, eig_vectors = eig(M_hat)
 
     # take sqrt to get the propagation constant
-    beta = np.sqrt(eig_values, dtype=np.complex128)
+    beta = np.sqrt(eig_values).astype(dtype)
     # invert eigenvalue if it corresponds to a backward propagating direction
     beta[np.real(beta) + np.imag(beta) < 0] *= -1
 
@@ -110,7 +111,7 @@ def fmm1d_te(lam, theta, period, perm_in, perm_out,
     # create grating
     G = (2 * np.pi / period) * np.arange(-N, N + 1, dtype=dtype)
     # create K_hat matrix
-    K_hat_square = np.diag((k_x + G) ** 2)
+    K_hat_square = np.diag((k_x + G) ** 2).astype(dtype)
     # initial phi electrical
     ident = np.identity(2 * N + 1, dtype=dtype)
 
@@ -123,13 +124,14 @@ def fmm1d_te(lam, theta, period, perm_in, perm_out,
     # iterate over all z layers
     for lt, perm in zip(layer_thicknesses, layer_perm):
         # get the betas and phi_e in this layer
-        beta, phi_e = fmm1d_te_layer_modes(perm, period, k_0, k_x, N)
+        beta, phi_e = fmm1d_te_layer_modes(perm, period, k_0, k_x, N,
+                                           dtype=dtype)
         # convert beta to beta_hat containing the entries on the diagonal
-        beta_hat = np.diag(beta)
+        beta_hat = np.diag(beta).astype(dtype)
 
         # matrices for forward and backward propagation
-        p_pos = np.diag(np.exp(1j * beta * lt))
-        p_neg = np.diag(np.exp(-1j * beta * lt))
+        p_pos = np.diag(np.exp(1j * beta * lt)).astype(dtype)
+        p_neg = np.diag(np.exp(-1j * beta * lt)).astype(dtype)
 
         # create A matrix which is needed to get the new transfer matrix
         A = np.block([[phi_e, phi_e],
@@ -143,11 +145,11 @@ def fmm1d_te(lam, theta, period, perm_in, perm_out,
 
     # beta_out_hat matrix
     beta_out_hat = np.sqrt(k_0 ** 2 * perm_out * ident
-                           - K_hat_square, dtype=np.complex128)
+                           - K_hat_square, dtype=dtype)
     beta_out_hat[np.real(beta_out_hat) + np.imag(beta_out_hat) < 0.0] *= -1
     # last missing matrix which inverse is left multiplied
     T_final = np.block([[ident, ident],
-                        [beta_out_hat, - beta_out_hat]])
+                        [beta_out_hat, - beta_out_hat]]).astype(dtype)
 
     # create the final transfer matrix
     T_matrix = solve(T_final, T_matrix)
@@ -171,7 +173,7 @@ def fmm1d_te(lam, theta, period, perm_in, perm_out,
     t = np.dot((t11 - t12 @ solve(t22, t21)), a_in[:, np.newaxis])
 
     # extract the diagonal elements
-    beta_in = np.diag(beta_0_hat)
+    beta_in = np.diag(beta_0_hat).astype(dtype)
 
     # calculate transmission and reflection efficiencies
     eta_r = np.array(np.real(1 / np.real(beta_in[N]) *
